+++ src/Symfony/Contracts/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add `HttpClientInterface::withOptions()`
+ * Add `TranslatorInterface::getLocale()`
 
 2.3.0
 -----